use BaseAWSLLM for bedrock getcredentials
